// src/generated/Protos/V1/Models/InferenceModels.ts
export enum ThinkLevels {
  ThinkLevelsBasic = 0,
  ThinkLevelsBasicWithTools = 1,
  ThinkLevelsSkilled = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface InferenceStatsResponse {
  SessionId: string;
  InferenceId: string;
  Success: boolean;
  LastMessageTokenCount: number;
  SessionTokenCount: number;
  LastMessageToolCount: number;
  SessionToolCount: number;
  LastMessageComputeTimeMs: number;
  SessionComputeTimeMs: number;
}